+++ gluecodium/src/test/resources/smoke/comments/output/android/com/example/smoke/Comments.java
@@ -41,27 +41,6 @@ public final class Comments extends NativeBase {
         public final Comments.SomeEnum error;
     }
     /**
-     * @exclude
-     */
-    static class SomeLambdaImpl extends NativeBase implements SomeLambda {
-        protected SomeLambdaImpl(final long nativeHandle, final Object dummy) {
-            super(nativeHandle, new Disposer() {
-                @Override
-                public void disposeNative(long handle) {
-                    disposeNativeHandle(handle);
-                }
-            });
-        }
-        private static native void disposeNativeHandle(long nativeHandle);
-        /**
-         * This is some very useful lambda that does it.
-         * @param p0 Very useful input parameter
-         * @param index Slightly less useful input parameter
-         * @return Usefulness of the input
-         */
-        public native double doIt(@NonNull final String p0, final int index);
-    }
-    /**
      * <p>This is some very useful struct.</p>
      */
     public static final class SomeStruct {
@@ -87,15 +66,36 @@ public final class Comments extends NativeBase {
         }
     }
     /**
-     * This is some very useful lambda that does it.
+     * @exclude
+     */
+    static class SomeLambdaImpl extends NativeBase implements SomeLambda {
+        protected SomeLambdaImpl(final long nativeHandle, final Object dummy) {
+            super(nativeHandle, new Disposer() {
+                @Override
+                public void disposeNative(long handle) {
+                    disposeNativeHandle(handle);
+                }
+            });
+        }
+        private static native void disposeNativeHandle(long nativeHandle);
+        /**
+         * <p>This is some very useful lambda that does it.</p>
+         * @param p0 <p>Very useful input parameter</p>
+         * @param index <p>Slightly less useful input parameter</p>
+         * @return <p>Usefulness of the input</p>
+         */
+        public native double doIt(@NonNull final String p0, final int index);
+    }
+    /**
+     * <p>This is some very useful lambda that does it.</p>
      */
     @FunctionalInterface
     public interface SomeLambda {
         /**
-         * This is some very useful lambda that does it.
-         * @param p0 Very useful input parameter
-         * @param index Slightly less useful input parameter
-         * @return Usefulness of the input
+         * <p>This is some very useful lambda that does it.</p>
+         * @param p0 <p>Very useful input parameter</p>
+         * @param index <p>Slightly less useful input parameter</p>
+         * @return <p>Usefulness of the input</p>
          */
         double doIt(@NonNull final String p0, final int index);
     }
@@ -183,7 +183,7 @@ public final class Comments extends NativeBase {
     public native boolean isSomeProperty();
     /**
      * <p>Sets some very useful property.</p>
-     * @param value Some very useful property.
+     * @param value <p>Some very useful property.</p>
      */
     public native void setSomeProperty(final boolean value);
 }